fix: add código da página no content e add sendResponse para finalizar evento
