+++ .env.example
@@ -1,1 +1,1 @@
-VITE_TAIGA_URL=
+VITE_TAIGA_URL=https://taiga.solarmarket.com.br/
