+++ public/manifest.json
@@ -22,8 +22,7 @@
       ],
       "js": [
         "content.js",
-        "popup.js",
-        "page.js"
+        "popup.js"
       ]
     }
   ],
